% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{add_wrap_id}
\alias{add_wrap_id}
\title{Add second ID for plotting.}
\usage{
add_wrap_id(obj, var = NULL)
}
\arguments{
\item{obj}{Object of class `stormtracks`}

\item{var}{The longitudes corresponding to which variables should be used. The default (NULL) uses the first column that starts on 'lon_'.}
}
\value{
The original object with an additional id (ID2) that changes when the stormtrack crosses the dateline
}
\description{
Adds a second ID to a stormtracks object that changes whenever a track crosses the dateline. This is useful to avoid plotting artifacts.
}
