% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{filter_by_duration}
\alias{filter_by_duration}
\title{Filter by duration}
\usage{
filter_by_duration(obj, dur_lim = c(0, Inf), unit = c("hours", "days"))
}
\arguments{
\item{obj}{Object of class 'stormtracks'}

\item{dur_lim}{Range of storm durations for which to return storms.}

\item{unit}{Unit in which duration is measured. Either `hours` or `days`. Can be abbreviated. Default is `hours`.}
}
\value{
Object of class 'stormtracks' containing only storms in the selected range of duration.
}
\description{
Filter storms by how long they lasted
}
\examples{
trx = read_tracks()
trx_1day = filter_by_duration(trx, dur_lim=c(1,1), unit='d')
}
