% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{filter_by_year}
\alias{filter_by_year}
\title{Filter by year}
\usage{
filter_by_year(obj, yr_lim = c(-Inf, Inf))
}
\arguments{
\item{obj}{Object of class 'stormtracks'}

\item{yr_lim}{Range of years for which to return storms}
}
\value{
Object of class 'stormtracks' containing only storms in the selected range of years.
}
\description{
Filter storms that occurred within a range of years.
}
\details{
The year range is inclusive, i.e. setting `yr_lim=c(1990,1991)` returns all storms that occurred in 1990 and 1991.
}
\examples{
trx = read_tracks()
trx_1990 = filter_by_year(trx, yr_lim=c(1990, 1990))
}
