% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{fortify.stormtracks}
\alias{fortify.stormtracks}
\title{Fortify storm tracks}
\usage{
\method{fortify}{stormtracks}(obj, ...)
}
\description{
Fortify storm tracks object into a data frame for plotting with ggplot2
}
