% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{plot.stormtracks}
\alias{plot.stormtracks}
\title{Plot storm tracks}
\usage{
plot.stormtracks(obj, ...)
}
\arguments{
\item{obj}{Object of class 'stormtrack'}

\item{...}{additional arguments passed to plot() function}
}
\description{
Plotting function for objects of class 'stormtrack'
}
\examples{
trx = read_tracks()
plot(trx, col='#00000010', lwd=5)
}
