% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{save_coastline}
\alias{save_coastline}
\title{create coast lines data frame from shapefile. this only has to be done once.}
\usage{
save_coastline()
}
\description{
create coast lines data frame from shapefile. this only has to be done once.
}
