% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{stormtracks_density}
\alias{stormtracks_density}
\title{Storm track density plot}
\usage{
stormtracks_density(obj, ...)
}
\arguments{
\item{obj}{Object of class 'stormtracks'}
}
\value{
A ggplot object.
}
\description{
Plot storm track density on a map.
}
\details{
This function takes all points in the stromtracks archive and constructs a 2d density plot using the ggplot2 function `stat_density2d`.
}
