% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{tidy.stormtracks}
\alias{tidy.stormtracks}
\title{Tidy storm tracks}
\usage{
\method{tidy}{stormtracks}(obj, ...)
}
\arguments{
\item{obj}{Object of class 'stormtracks'}
}
\value{
A tidy data frame.
}
\description{
Tidy up storm tracks into a data frame for use with ggplot2.
}
