% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_density.R
\name{track_density}
\alias{track_density}
\title{Calculate storm track density on a grid.}
\usage{
track_density(tracks, bw = 1, var = NULL)
}
\arguments{
\item{tracks}{Object of class `stormtracks` (see `read_tracks`)}

\item{bw}{The bin width (in degrees) of the grid on which track density is calculate. Default=1.}

\item{var}{Which variable to use to define latitude and longitude. If `NULL`
(the default) it uses the first column names that start on `lon_` and `lat_`.}
}
\value{
grid of counts
}
\description{
Count total number and normalised number of storm tracks or features on a grid.
}
\examples{
tracks = read_tracks('../inst/extdata/storms.txt')
counts = count_storms(tracks)
}
\seealso{
read_tracks plot.track_density
}
