% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{unwrap_lonlat}
\alias{unwrap_lonlat}
\title{Unwrap longitudes and latitudes}
\usage{
unwrap_lonlat(lonlat)
}
\arguments{
\item{lonlat}{A 2 column matrix, each row is a lon/lat pair.}
}
\value{
A list of 2-column matrices lon/lat.
}
\description{
Unwrap longitudes to plot lines properly on a 2d map
}
\details{
This is a helper function used in plot.stormtracks below. When
longitude crosses the date line, the line has to be split into two
segments. Otherwise R will connect the line segments across the entire
globe. The built-in R function maptools::nowrapSpatialLines seems to
be broken, so I coded it by hand. This function splits a 2 column
matrix of lons and lats into a list of 2-column matrices matrices,
separated whenever lon crosses the dateline.
}
