% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tracks.R
\name{plot.stormtracks}
\alias{plot.stormtracks}
\title{Plot storm tracks}
\usage{
plot.stormtracks(obj, var = NULL)
}
\arguments{
\item{obj}{Object of class 'stormtracks'}

\item{var}{Which variable to use for plotting. The default (NULL) uses the first columns in the data table that start on 'lon_', 'lat_' and 'var_'.}
}
\value{
A ggplot object.
}
\description{
Plotting function for objects of class 'stormtracks'
}
