% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_tracks.R
\name{read_tracks}
\alias{read_tracks}
\title{Read storm tracks}
\usage{
read_tracks(filename)
}
\arguments{
\item{filename}{Input file name}
}
\value{
An object of class 'stormtracks'
}
\description{
Create a list of storm tracks from a file.
}
\details{
A stormtracks-object is a list of storm tracks. Each storm track is represented by a matrix with columns (at least) dates, lat, lon, vorticity, and possibly further variables like sea level pressure or precipitation.
}
