% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{stormtracks_density}
\alias{stormtracks_density}
\title{Storm track density plot}
\usage{
stormtracks_density(obj, type = c("density_2d", "bin_2d"), ...)
}
\arguments{
\item{obj}{Object of class 'stormtracks'}

\item{type}{Which type of density estimator to use. Options are `density_2d` (the default) and `bin_2d`. See the help files of `ggplot2::stat_bin_2d` and `ggplot2::stat_density_2d` for details.}
}
\value{
A ggplot object.
}
\description{
Deprecated. Use the plot() function for objects of type `track_density` instead.
}
\details{
This function takes all points in the stromtracks archive and constructs a 2d density plot using the ggplot2 function `stat_density2d`.
}
